refactor: Handler 인터페이스화
